Assignment-1/requirements.txt
nltk
tensorflow
python-docx
youtube-search-python
youtube_transcript_api
